refactor(components): improve imports | hj | #3176
