+++ packages/components/src/components/flyout/flyout.tsx
@@ -24,7 +24,7 @@ import {
 } from '../../utils';
 import type { PropTypes, SelectedAriaAttributes, Theme } from '../../types';
 import { throttle } from 'throttle-debounce';
-import { getSlottedAnchorStyles } from '../../styles/global/slotted-anchor-styles';
+import { getSlottedAnchorStyles } from '../../styles';
 
 const propTypes: PropTypes<typeof Flyout> = {
   open: AllowedTypes.boolean,
