+++ packages/components/src/components/table/table/table.tsx
@@ -12,7 +12,7 @@ import type { PropTypes, Theme } from '../../../types';
 import { getComponentCss } from './table-styles';
 import type { TableUpdateEventDetail } from './table-utils';
 import { SORT_EVENT_NAME, warnIfCaptionIsMissing } from './table-utils';
-import { getSlottedAnchorStyles } from '../../../styles/global/slotted-anchor-styles';
+import { getSlottedAnchorStyles } from '../../../styles';
 
 const propTypes: PropTypes<typeof Table> = {
   caption: AllowedTypes.string,
