+++ packages/components/src/components/text-field-wrapper/text-field-wrapper.tsx
@@ -51,7 +51,7 @@ import {
   UNIT_POSITIONS,
 } from './text-field-wrapper-utils';
 import { Label } from '../common/label/label';
-import { getSlottedAnchorStyles } from '../../styles/global/slotted-anchor-styles';
+import { getSlottedAnchorStyles } from '../../styles';
 import { getSlottedInputIndicatorStyles } from '../../styles/global/slotted-input-indicator-styles';
 
 const propTypes: PropTypes<typeof TextFieldWrapper> = {
