+++ packages/components/src/styles/global/index.ts
@@ -1,0 +1,3 @@
+// Since slotted input indicator is only relevant for one component we don't want it to bundle into main chunk
+export * from './slotted-anchor-styles';
+export * from './slotted-picture-image-styles';
